get-orig-source: print the list of directories vendored by upstream
This script is a good place to do that, in the sense that we have access
to the un-modified upstream archive at this point.

The idea is to have the list of vendoreded dependencies, in a format
suitable for d/copyright. So that when we attempt to package a new version
of Docker, we can update the Files-Excluded field list without effort.

By keeping this list up to date, we can have an instant overview of the
vendoring status, and hence a bit of a todo list. At the moment, some
of the vendored directories were introduced by version 18.06, and I
still didn't have time to check what needs to be packaged.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>